decontainer setup

+++ hooks/post_gen_project.py
@@ -45,6 +45,13 @@ def remove_gplv3_files():
         os.remove(file_name)
 
 
+def remove_vscode_devcontainer_files():
+    dir_paths = [".devcontainer", ".history"]
+    for dir_path in dir_paths:
+        if os.path.exists(dir_path):
+            shutil.rmtree(dir_path)
+
+
 def remove_pycharm_files():
     idea_dir_path = ".idea"
     if os.path.exists(idea_dir_path):
@@ -56,11 +63,13 @@ def remove_pycharm_files():
 
 
 def remove_docker_files():
-    shutil.rmtree("compose")
+    if "{{ cookiecutter.use_vscode_devcontainer }}".lower() == "n":
+        shutil.rmtree("compose")
 
-    file_names = ["local.yml", "production.yml", ".dockerignore"]
-    for file_name in file_names:
-        os.remove(file_name)
+        file_names = ["local.yml", "production.yml", ".dockerignore"]
+        for file_name in file_names:
+            os.remove(file_name)
+
     if "{{ cookiecutter.use_pycharm }}".lower() == "y":
         file_names = ["docker_compose_up_django.xml", "docker_compose_up_docs.xml"]
         for file_name in file_names:
@@ -448,6 +457,9 @@ def main():
         remove_utility_files()
     else:
         remove_docker_files()
+
+    if "{{ cookiecutter.use_vscode_devcontainer }}".lower() == "n":
+        remove_vscode_devcontainer_files()
 
     if (
         "{{ cookiecutter.use_docker }}".lower() == "y"
